@@ -53,14 +53,14 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.0.8",
-    "@angular/common": "^17.0.8",
-    "@angular/compiler": "^17.0.8",
-    "@angular/core": "^17.0.8",
-    "@angular/forms": "^17.0.8",
-    "@angular/platform-browser": "^17.0.8",
-    "@angular/platform-browser-dynamic": "^17.0.8",
-    "@angular/router": "^17.0.8",
+    "@angular/animations": "^18.0.1",
+    "@angular/common": "^18.0.1",
+    "@angular/compiler": "^18.0.1",
+    "@angular/core": "^18.0.1",
+    "@angular/forms": "^18.0.1",
+    "@angular/platform-browser": "^18.0.1",
+    "@angular/platform-browser-dynamic": "^18.0.1",
+    "@angular/router": "^18.0.1",
     "@schematics/angular": "^16.0.2",
     "rxjs": "^7.4.0",
     "shelljs": "^0.8.5",
@@ -68,24 +68,24 @@
     "zone.js": "~0.14.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.0.8",
-    "@angular-devkit/core": "^17.0.8",
-    "@angular-devkit/schematics": "^17.0.8",
-    "@angular-eslint/builder": "^17.1.1",
-    "@angular-eslint/eslint-plugin": "^17.1.1",
-    "@angular-eslint/eslint-plugin-template": "^17.1.1",
-    "@angular-eslint/schematics": "^17.1.1",
-    "@angular-eslint/template-parser": "^17.1.1",
-    "@angular/cli": "^17.0.8",
-    "@angular/compiler-cli": "^17.0.8",
+    "@angular-devkit/build-angular": "^18.0.2",
+    "@angular-devkit/core": "^18.0.2",
+    "@angular-devkit/schematics": "^18.0.2",
+    "@angular-eslint/builder": "^18.0.0",
+    "@angular-eslint/eslint-plugin": "^18.0.0",
+    "@angular-eslint/eslint-plugin-template": "^18.0.0",
+    "@angular-eslint/schematics": "^18.0.0",
+    "@angular-eslint/template-parser": "^18.0.0",
+    "@angular/cli": "^18.0.2",
+    "@angular/compiler-cli": "^18.0.1",
     "@evilmartians/lefthook": "^1.0.3",
     "@types/jasmine": "^4.0.0",
     "@types/node": "^16.11.7",
-    "@typescript-eslint/eslint-plugin": "^6.10.0",
-    "@typescript-eslint/parser": "^6.10.0",
+    "@typescript-eslint/eslint-plugin": "^7.2.0",
+    "@typescript-eslint/parser": "^7.2.0",
     "copyfiles": "^2.4.1",
     "coveralls": "^3.1.0",
-    "eslint": "^8.53.0",
+    "eslint": "^8.57.0",
     "eslint-plugin-import": "2.27.5",
     "eslint-plugin-jsdoc": "44.2.4",
     "eslint-plugin-prefer-arrow": "1.2.3",
@@ -97,10 +97,10 @@
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "^2.0.0",
-    "ng-packagr": "^17.0.3",
+    "ng-packagr": "^18.0.0",
     "prettier": "^2.2.1",
     "rfc4648": "^1.5.0",
     "ts-node": "^10.9.1",
-    "typescript": "~5.2.0"
+    "typescript": "~5.4.5"
   }
 }
