@@ -1,14 +1,11 @@
 import { CommonModule } from '@angular/common';
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { ModuleWithProviders, NgModule } from '@angular/core';
 import { PassedInitialConfig } from './auth-config';
 import { _provideAuth } from './provide-auth';
 
-@NgModule({
-  imports: [CommonModule, HttpClientModule],
-  declarations: [],
-  exports: [],
-})
+@NgModule({ declarations: [],
+    exports: [], imports: [CommonModule], providers: [provideHttpClient(withInterceptorsFromDi())] })
 export class AuthModule {
   static forRoot(
     passedConfig: PassedInitialConfig
