@@ -1,4 +1,4 @@
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { AppComponent } from './app.component';
@@ -9,16 +9,12 @@ import { HomeComponent } from './home/home.component';
 import { NavigationComponent } from './navigation/navigation.component';
 import { UnauthorizedComponent } from './unauthorized/unauthorized.component';
 
-@NgModule({
-  imports: [BrowserModule, routing, HttpClientModule, AuthConfigModule],
-  declarations: [
-    AppComponent,
-    ForbiddenComponent,
-    HomeComponent,
-    NavigationComponent,
-    UnauthorizedComponent,
-  ],
-  providers: [],
-  bootstrap: [AppComponent],
-})
+@NgModule({ declarations: [
+        AppComponent,
+        ForbiddenComponent,
+        HomeComponent,
+        NavigationComponent,
+        UnauthorizedComponent,
+    ],
+    bootstrap: [AppComponent], imports: [BrowserModule, routing, AuthConfigModule], providers: [provideHttpClient(withInterceptorsFromDi())] })
 export class AppModule {}
