@@ -1,4 +1,4 @@
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { BrowserModule } from '@angular/platform-browser';
@@ -11,22 +11,16 @@ import { HomeComponent } from './home/home.component';
 import { NavigationComponent } from './navigation/navigation.component';
 import { UnauthorizedComponent } from './unauthorized/unauthorized.component';
 
-@NgModule({
-  imports: [
-    BrowserModule,
-    FormsModule,
-    routing,
-    HttpClientModule,
-    AuthConfigModule,
-  ],
-  declarations: [
-    AppComponent,
-    ForbiddenComponent,
-    HomeComponent,
-    AutoLoginComponent,
-    NavigationComponent,
-    UnauthorizedComponent,
-  ],
-  bootstrap: [AppComponent],
-})
+@NgModule({ declarations: [
+        AppComponent,
+        ForbiddenComponent,
+        HomeComponent,
+        AutoLoginComponent,
+        NavigationComponent,
+        UnauthorizedComponent,
+    ],
+    bootstrap: [AppComponent], imports: [BrowserModule,
+        FormsModule,
+        routing,
+        AuthConfigModule], providers: [provideHttpClient(withInterceptorsFromDi())] })
 export class AppModule {}
